- Добавлен STM32F103 для интерфейса управления - Подготовка к реализации motion controller - Иготовлен корпус для электроники draft

+++ firmware/EDM/src/spark.h
@@ -1,6 +1,6 @@
 #ifndef __SPARK_H__
 #define __SPARK_H__
-#include "Arduino.h"
+#include "core.h"
 
 
 void spark_pwm_init();
@@ -9,11 +9,11 @@ void spark_pwm_stop();
 void spark_pwm_update();
 void spark_pwm_process();
 
-void spark_set_t1_us(int32_t t);
-void spark_set_t0_us(int32_t t);
-int32_t spark_get_t1_us();
-int32_t spark_get_t0_us();
-int32_t spark_get_freq();
+void spark_set_t1_us(uint16_t t);
+void spark_set_t0_us(uint16_t t);
+uint16_t spark_get_t1_us();
+uint16_t spark_get_t0_us();
+uint16_t spark_get_freq();
 
 bool spark_is_enabled();
